% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_distortion.R
\name{circle_distortion}
\alias{circle_distortion}
\title{circle_distortion}
\usage{
circle_distortion(
  n = 1000,
  X_rate = 2,
  Y_rate = 2,
  time_sampling = "unif",
  nb_levels = 20
)
}
\arguments{
\item{n}{Overall number of points to draw}

\item{X_rate}{x axis distortion over time}

\item{Y_rate}{y axis distortion over time}

\item{time_sampling}{unif (uniform continuous time sampling) or discrete (data are located at discrete regular time steps)}

\item{nb_levels}{number of time steps (for option sampling = "discrete")}
}
\value{
tibble with 3 columns (t,x,y) and n rows
}
\description{
Function generating a tibble with 3 columns named (t,x,y) with data in the (x,y) plane forming a circle transformed into an ellipse over time. (Using the right distance between Persistence Diagrams, this example should be change-point free)
}
\examples{
circle_distortion()
}
