% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_extinction.R
\name{circle_extinction}
\alias{circle_extinction}
\title{circle_extinction}
\usage{
circle_extinction(
  n = 1000,
  change = 0.5,
  time_sampling = "unif",
  nb_levels = 20
)
}
\arguments{
\item{n}{Overall number of points to draw}

\item{change}{Relative location of the change (example: \code{change = 0.5} = at the middle of the data, time goes from 0 to 1 with this function)}

\item{time_sampling}{\code{"unif"} (uniform continuous time sampling) or \code{"discrete"} (data are located at discrete regular time steps)}

\item{nb_levels}{number of time steps (for option \code{sampling = "discrete"}). The n points are distributed uniformily (same amount by level when possible) on all the \code{nb_levels}.}
}
\value{
tibble with 3 columns (t,x,y) and n rows
}
\description{
Function generating a tibble with 3 columns named (t,x,y) with data in the (x,y) plane forming a circle transformed into a segment by reduction of one of the semi-axis
}
\examples{
circle_extinction()
}
