% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_move_distortion.R
\name{circle_move_distortion}
\alias{circle_move_distortion}
\title{circle_move_distortion}
\usage{
circle_move_distortion(
  n = 1000,
  final_position = c(2, 3),
  rotation = 0.5,
  time_sampling = "unif",
  nb_levels = 20,
  X_rate = 2,
  Y_rate = 0.3
)
}
\arguments{
\item{n}{Overall number of points to draw}

\item{final_position}{the x-y coordinate of the barycenter of the final ellipse at \code{time = 1}. The figure moves linearly with time to this position.}

\item{rotation}{the number of rotations of the final figure, which is rotated linearly with time (given in radians as \code{rotation x 2 pi}). If equal to 0.5, the rotation angle is pi for example.}

\item{time_sampling}{\code{"unif"} (uniform continuous time sampling) or \code{"discrete"} (data are located at discrete regular time steps)}

\item{nb_levels}{number of time steps (for option \code{sampling = "discrete"}). The n points are distributed uniformily (same amount by level when possible) on all the \code{nb_levels}.}

\item{X_rate}{regular-in-time multiplicative factor for the x axis (x-coordinates are multiplicated by \code{(X_rate*time + 1 - time)} where \code{time} goes from 0 to 1)}

\item{Y_rate}{regular-in-time multiplicative factor for the y axis (y-coordinates are multiplicated by \code{(Y_rate*time + 1 - time)} where \code{time} goes from 0 to 1)}
}
\value{
tibble with 3 columns (t,x,y) and n rows
}
\description{
Function generating a tibble with 3 columns named (t,x,y) with data in the (x,y) plane forming a circle transformed into an ellipse over time with rotation and shift of the center position. (Using the right distance between Persistence Diagrams, this example should be change-point free)
}
\examples{
circle_move_distortion()
}
