% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2D_pointCircleGap.R
\name{data2D_pointCircleGap}
\alias{data2D_pointCircleGap}
\title{data2D_pointCircleGap}
\usage{
data2D_pointCircleGap(n, gap)
}
\arguments{
\item{n}{Overall number of points to draw}

\item{gap}{size of the gap}
}
\value{
tibble with 2 columns (x,y) and n rows
}
\description{
Function generating a tibble with 2 columns named (x,y) with 2D data forming on a circle with gap
}
\examples{
data2D_pointCircleGap(1000, 0.1)
}
