% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2D_pointEllipseMissingArc.R
\name{data2D_pointEllipseMissingArc}
\alias{data2D_pointEllipseMissingArc}
\title{data2D_pointEllipseMissingArc}
\usage{
data2D_pointEllipseMissingArc(
  n = 1000,
  gap = 0.1,
  sdNoise = 0.1,
  a = 1,
  b = 1,
  gapStart = 0
)
}
\arguments{
\item{n}{Overall number of points to draw (around the circle of radius 1)}

\item{gap}{size of the missing arc as a percent of the global arc length (2 pi)}

\item{sdNoise}{additional Gaussian noise}

\item{a}{coefficient in front of cos(t)}

\item{b}{coefficient in front of sin(t)}

\item{gapStart}{uniform values between gapStart and 2*pi - gap + gapStart for the uniform t angles}
}
\value{
tibble with 2 columns (x,y) and n rows
}
\description{
Function generating a tibble with 2 columns named (x,y) with 2D data forming on a circle with gap (a missing arc)
}
\examples{
data2D_pointEllipseMissingArc(1000, 0.1)
}
