% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2D_pointSquareHole.R
\name{data2D_pointSquareHole}
\alias{data2D_pointSquareHole}
\title{data2D_pointSquareHole}
\usage{
data2D_pointSquareHole(n, Hole_Relative_length = 0)
}
\arguments{
\item{n}{Overall number of points to draw}

\item{Hole_Relative_length}{size of the hole in the center of the square}
}
\value{
tibble with 2 columns (x,y) and n rows
}
\description{
Function generating a tibble with 2 columns named (x,y) with 2D data forming in a 1by1 square with a hole in its center of size \code{Hole_Relative_length}
}
\examples{
data2D_pointSquareHole(1000)
}
