% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2D_pointTwoCirclesMerged.R
\name{data2D_pointTwoCirclesMerged}
\alias{data2D_pointTwoCirclesMerged}
\title{data2D_pointTwoCirclesMerged}
\usage{
data2D_pointTwoCirclesMerged(n = 1000, overlap = 0.1, sdNoise = 0.05)
}
\arguments{
\item{n}{Overall number of points to draw (two circle of radius 1 with overlap)}

\item{overlap}{we have two circle of radius 1, the first located in (0,0), the second in (0, 2-overlap)}

\item{sdNoise}{additional Gaussian noise}
}
\value{
tibble with 2 columns (x,y) and n rows
}
\description{
Function generating a tibble with 2 columns named (x,y) with 2D data on two overlapping circles (no point in the overlap)
}
\examples{
data2D_pointTwoCirclesMerged(1000, 0.1)
}
