% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data3D_almostCircle_to_circle.R
\name{data3D_almostCircle_to_circle}
\alias{data3D_almostCircle_to_circle}
\title{data3D_almostCircle_to_circle}
\usage{
data3D_almostCircle_to_circle(
  n = 1000,
  gap = 0.1,
  change = 0.5,
  time_sampling = "unif",
  nb_levels = 20,
  X_rate = 1,
  Y_rate = 1
)
}
\arguments{
\item{n}{Overall number of points to draw}

\item{gap}{distance between the circle and the segment From 0 to 1. 0 = the circle, 1 = the segment}

\item{change}{Relative location of the change (example: \code{change = 0.5} = at the middle of the data, time goes from 0 to 1 with this function)}

\item{time_sampling}{\code{"unif"} (uniform continuous time sampling) or \code{"discrete"} (data are located at discrete regular time steps)}

\item{nb_levels}{number of time steps (for option \code{sampling = "discrete"}). The n points are distributed uniformily (same amount by level when possible) on all the \code{nb_levels}.}

\item{X_rate}{regular-in-time multiplicative factor for the x axis (x-coordinates are multiplicated by \code{(X_rate*time + 1 - time)} where \code{time} goes from 0 to 1)}

\item{Y_rate}{regular-in-time multiplicative factor for the y axis (y-coordinates are multiplicated by \code{(Y_rate*time + 1 - time)} where \code{time} goes from 0 to 1)}
}
\value{
tibble with 3 columns (t,x,y) and n rows
}
\description{
Function generating a tibble with 3 columns named (t,x,y) with data in the (x,y) plane forming a circle almost closed, closing at time step \cite{change} into a circle. The transformation is "isometric" as the length of the object stays the same over time (= 2 x pi) (only if no distortion with \code{X_Rate = 1} and \code{Y_Rate = 1})
}
\examples{
data3D_almostCircle_to_circle()
}
