% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EulerCurve_Uniform.R
\name{dataUniform}
\alias{dataUniform}
\title{dataUniform}
\usage{
dataUniform(N, L1 = 1, L2 = 1, M = 1, a = 1, b = 1)
}
\arguments{
\item{N}{number of points to generate}

\item{L1}{segment length, interval 1}

\item{L2}{segment length, interval 2}

\item{M}{distance between the intervals}

\item{a}{density level for interval 1}

\item{b}{density level for interval 2}
}
\description{
Uniform data over 2 segments with segment lengths L1,L2, constant densities a, b and N data points
}
\examples{
dataUniform(1000)
}
