% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{distances_Persistence_Diagrams}
\alias{distances_Persistence_Diagrams}
\title{distances_Persistence_Diagrams}
\usage{
distances_Persistence_Diagrams(
  PD,
  distance = "wasserstein",
  type = "2by2",
  dimension = 1
)
}
\arguments{
\item{PD}{Persistence diagram with 4 columns:  time, dimension, Birth, Death.}

\item{distance}{Type of distance used}

\item{type}{1by1 (segment versus segment) comparisons or 2by2 (PD by PD) comparisons}

\item{dimension}{dimension to chose in Persitence diagram}
}
\description{
Function computing distances between Persistence Diagrams
}
\examples{
data <- segment_to_circle(n = 1000, time_sampling = "discrete", nb_levels = 20)
all_PD <- Generate_All_Persistence_Diagrams(data, nb_levels = 20)
distances_Persistence_Diagrams(all_PD)
}
