% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EulerCurve_Uniform.R
\name{EulerCPC_exact}
\alias{EulerCPC_exact}
\title{EulerCPC_exact}
\usage{
EulerCPC_exact(r, N, N1, N2, L1, L2, M)
}
\arguments{
\item{r}{evaluation point (radius r)}

\item{N}{number of points}

\item{N1}{number of points in segment 1}

\item{N2}{number of points in segment 2}

\item{L1}{segment length, interval 1}

\item{L2}{segment length, interval 2}

\item{M}{distance between the intervals}
}
\description{
Exact value of the Euler curve for 1D Uniform data (with 2 segments)
}
