% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EulerCurve_Uniform.R
\name{EulerCPC_VAR_approx}
\alias{EulerCPC_VAR_approx}
\title{EulerCPC_VAR_approx}
\usage{
EulerCPC_VAR_approx(r, N, L)
}
\arguments{
\item{r}{evaluation point (radius r)}

\item{N}{number of points}

\item{L}{segment length}
}
\description{
Approximate value of the Variance of the Euler curve for 1D Uniform data (with 1 segment)
}
