% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersistenceDiagrams_TS.R
\name{Generate_All_Persistence_Diagrams}
\alias{Generate_All_Persistence_Diagrams}
\title{Generate_All_Persistence_Diagrams}
\usage{
Generate_All_Persistence_Diagrams(data, nb_levels = 20)
}
\arguments{
\item{data}{tibble data with columns (time,x,y)}

\item{nb_levels}{number of regular slices to be considered for the time variable}
}
\value{
tibble of the sequence of Persistence Diagrams with columns (time, dimension, Birth, Death). "time" is the position of the persistence diagram in the sequence of diagrams.
}
\description{
Function generating a sequence of Persistence Diagrams using tibble data with columns (time,x,y)
}
