% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersistenceDiagram_operations.R
\name{proj_Points}
\alias{proj_Points}
\title{proj_Points}
\usage{
proj_Points(PD, v = 0, h = 0)
}
\arguments{
\item{PD}{Persistence diagram with 4 columns:  time, dimension, Birth, Death.}

\item{v}{vertical threshold}

\item{h}{horizontal threshold}
}
\value{
Points of the PD projected
}
\description{
Generating projective points from the persistence diagram in the diagonal, a vertical line and an horizontal line
}
