% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersistentDiagram_operations.R
\name{remove_noisy_points}
\alias{remove_noisy_points}
\title{remove_noisy_points}
\usage{
remove_noisy_points(PD, v = 0, h = 0, infinity = FALSE)
}
\arguments{
\item{PD}{Persistence diagram with 4 columns:  time, dimension, Birth, Death.}

\item{v}{vertical threshold}

\item{h}{horizontal threshold}

\item{infinity}{if FALSE we remove points with infinite "Death" value}
}
\value{
The PD filtrated by the 2 boundaries
}
\description{
Removing points from the persistence diagram out of the vertical and horizontal boundaries
}
