% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_to_circle.R
\name{segment_to_circle}
\alias{segment_to_circle}
\title{segment_to_circle}
\usage{
segment_to_circle(
  n = 1000,
  change = 0.5,
  time_sampling = "unif",
  nb_levels = 20
)
}
\arguments{
\item{n}{Overall number of points to draw}

\item{change}{Relative location of the change (example: 0.5 = at the middle of the data, time goes from 0 to 1 with this function)}

\item{time_sampling}{unif (uniform continuous time sampling) or discrete (data are located at discrete regular time steps)}

\item{nb_levels}{number of time steps (for option sampling = "discrete")}
}
\value{
tibble with 3 columns (t,x,y) and n rows
}
\description{
Function generating a tibble with 3 columns named (t,x,y) with data in the (x,y) plane forming a segment closing to a circle at the change-point location.
}
\examples{
segment_to_circle()
}
