% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersistentDiagram_operations.R
\name{Generate_All_Persistence_diagrams}
\alias{Generate_All_Persistence_diagrams}
\title{Generate_All_Persistence_diagrams}
\usage{
Generate_All_Persistence_diagrams(data, nbLevel = 20)
}
\value{
Points of the PD projected
}
\description{
Generate_All_Persistence_diagrams
}
