% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersistenceDiagram_operations.R
\name{remove_noisy_points}
\alias{remove_noisy_points}
\title{remove_noisy_points}
\usage{
remove_noisy_points(PD, birth = Inf, death = 0, infinity = FALSE)
}
\arguments{
\item{PD}{Persistence diagram with 4 columns:  time, dimension, Birth, Death.}

\item{birth}{vertical threshold}

\item{death}{horizontal threshold}

\item{infinity}{if FALSE we remove points with infinite "Death" value}
}
\value{
The PD filtrated by the 2 boundaries (and the Inf points)
}
\description{
Removing points from the persistence diagram out of the vertical and horizontal boundaries
}
