% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_to_circle.R
\name{segment_to_circle}
\alias{segment_to_circle}
\title{segment_to_circle}
\usage{
segment_to_circle(nb = 1000, change = 0.5, sampling = "unif", level = 20)
}
\arguments{
\item{nb}{Overall number of points}

\item{change}{The relative location of the change (0.5 = at the middle of the data)}

\item{sampling}{unif or regular}

\item{level}{number of time steps}
}
\value{
A data frame with 3 columns and nb rows
}
\description{
Generating a data frame with 3 columns (x,y,t)
}
\examples{
segment_to_circle()
}
